Convert `TransactionSubscriber` to Listeners
